Add Array3D::restore_from()

+++ types.h
@@ -3,10 +3,9 @@
 
 #include <complex>
 
-// In principle, this gives flexibility, but one also would need to adjust the
-// FFTW calls.
+// Our in-place FFTs require sizeof(Complex) = 2*sizeof(Float).
 typedef double Float;
-typedef std::complex<double> Complex;
+typedef std::complex<Float> Complex;
 
 // We want to allow that the FFT grid could exceed 2^31 cells
 typedef unsigned long long int uint64;
